pyproject.toml
[project]
name = "agentos-tools"
version = "0.1.0"
description = "AgentOS configuration and tooling"
authors = [{name = "Marty McEnroe"}]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pywinpty (>=3.0.2,<4.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "wexpect (>=4.0.0,<5.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "keyring (>=25.7.0,<26.0.0)",
    "anthropic (>=0.76.0,<0.77.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]